requirements.txt
newspaper3k
pytz